Use pkg-config to determine json-c header path
When cross compiling, the system include directory will not necessarily
be /usr/include, so /usr/include/json-c may not be the correct location
for the json-c headers. Using pkg-config allows the sysroot to be
specified at build time, which fixes this.

This patch updates the src/Makefile to use pkg-config to locate the
json-c headers, and updates the packaging to include pkg-config as a
build dependency.